requirements.txt
streamlit
google-genai
feedparser
python-dotenv
requests
mcp
openmeteo-requests